feat: gameplay
- save: balance, business components
- upgrade income
- buy other business
- logic level up button

+++ Assets/_Project/Scripts/_Configs/PriceLevelUpConfig.cs
@@ -5,6 +5,6 @@ namespace _Project.Scripts._Configs
     [CreateAssetMenu(fileName = "PriceLVLUp", menuName = "Config/LVLUp", order = 51)]
     public class PriceLevelUpConfig : ScriptableObject
     {
-        [field: SerializeField] public int[] Price { get; private set; }
+        [field: SerializeField] public double[] Price { get; private set; }
     }
 }
